Add code for adding metadata to iiif manifest to all work types, not just paged resources.
